add type room. when message resended, field byuser set by original message.

+++ server/src/chat/block-messages/block-messages.service.ts
@@ -6,13 +6,18 @@ import { Repository } from 'typeorm';
 import { MessageEntity } from '../message/message.entity';
 import { RoomEntity } from '../room/room.entity';
 import { MessageService } from '../message/message.service';
+import { NotifyService } from '../notify/notify.service';
+import { NotifyEntity } from '../notify/notify.entity';
 
 @Injectable()
 export class BlockMessagesService {
   constructor(
     @InjectRepository(BlockMessagesEntity)
     private readonly blockMessagesRepository: Repository<BlockMessagesEntity>,
+    @InjectRepository(NotifyEntity)
+    private readonly notifyRepository: Repository<NotifyEntity>,
     private messageService: MessageService,
+    private notifyService: NotifyService,
   ) {}
 
   async createBlock(room: RoomEntity) {
@@ -40,7 +45,12 @@ export class BlockMessagesService {
       where: {
         id,
       },
-      relations: { messages: true },
+      relations: {
+        messages: true,
+        notifies: {
+          afterMessage: true,
+        },
+      },
     });
   }
 
@@ -58,12 +68,17 @@ export class BlockMessagesService {
     };
   }
 
-  async resendMessage(message: MessageEntity, room: RoomEntity) {
+  async resendMessage(
+    message: MessageEntity,
+    room: RoomEntity,
+    user: UserEntity,
+  ) {
     const { block, isNewBlock } = await this.checkExistBlock(room);
 
     const resendingMessage = await this.messageService.resendMessage(
       message,
       block,
+      user,
     );
 
     block.messages.push(resendingMessage);
@@ -109,14 +124,64 @@ export class BlockMessagesService {
 
   async checkIsEmptyBlock(id: number) {
     const block = await this.findById(id);
+
     let idBlock: number;
-    // TODO: Проверка и на пустоту оповещения (notifies)
-    if (!block.messages.length) {
+    if (!block.messages.length && !block.notifies.length) {
       idBlock = block.id;
       await this.deleteBlock(block);
     }
 
     return idBlock;
+  }
+
+  async deleteMessage(message: MessageEntity) {
+    const block = await this.findById(message.block.id);
+
+    const notifies = block.notifies.filter(
+      (n) => n.afterMessage?.id === message.id,
+    );
+
+    if (notifies.length) {
+      message.notifies = [];
+
+      let index = block.messages.findIndex((m) => m.id === message.id);
+
+      if (index >= 0) {
+        index--;
+      }
+
+      const prevMessage = block.messages[index];
+
+      notifies.forEach(async (n) => {
+        n.afterMessage = prevMessage || null;
+        await this.notifyRepository.save(n);
+      });
+    }
+
+    await this.messageService.deleteMessage(message);
+
+    const idBlock = this.checkIsEmptyBlock(block.id);
+
+    return idBlock;
+  }
+
+  async createNotify(room: RoomEntity, body: string, user: UserEntity) {
+    const { block, isNewBlock } = await this.checkExistBlock(room);
+
+    const message = block.messages.at(-1);
+
+    const notify = await this.notifyService.createNotify(
+      body,
+      user,
+      message,
+      block,
+    );
+
+    return {
+      ...this.notifyService.getNotifyFields(notify),
+      idRoom: room.id,
+      block: isNewBlock ? this.getBlockFields(block) : undefined,
+    };
   }
 
   getBlockFields(block: BlockMessagesEntity) {
